Fixed bugs from test cases. Also, organized files better
