--- Project-5-Space-Craft/Class1.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Project_5_Space_Craft
-{
-    internal class Class1
-    {
-    }
-}
