+++ Project-5-Space-Craft/Functions/FunctionStub.cs
@@ -1,7 +1,7 @@
 ﻿//PayloadOps TODO:// Also remove this stub
 //This is an example stub for an implementation of Functions and commands
 
-namespace Project_5_Space_Craft
+namespace Project_5_Space_Craft.Functions
 {
     public class FunctionStub : IFunction
     {
@@ -14,7 +14,7 @@ namespace Project_5_Space_Craft
 
         public string GetCommand()
         {
-            return "Some Command";
+            return this.command;
         }
 
         public void RunCommand()
