+++ Project-5-Space-Craft/Readings/ReadingsStub.cs
@@ -1,12 +1,13 @@
 ﻿//PayloadOps TODO:// Remove this stub
 //This is an example stud for an implementation of SpaceshipReadings
-namespace Project_5_Space_Craft
+namespace Project_5_Space_Craft.Readings
 {
     public class ReadingsStub : IReading
     {
         private string data;
 
-        public ReadingsStub(string data) {
+        public ReadingsStub(string data)
+        {
             this.data = data;
         }
